TST #823 continue generalizing
